Refactor PngImage to consistently use ArrayBuffer
Code refactor to carefully use ArrayBuffer instead of ByteArray to
leverage DataViews and other constructs. The png raw data is stored in
an ArrayBuffer. Changes through the interface change this ArrayBuffer.
The GUI elements should update to reflect this changes using listeners.
The edited image is also updated (by means of a listener).